fix publisher
