+++ SuscriberService/entrypoint.prod.sh
