+++ SuscriberService/requirements.txt
@@ -1,0 +1,23 @@
+celery==5.2.7
+flask==2.2.2
+Flask-JWT==0.2.0
+flask_jwt_extended==4.4.4
+Flask-RESTful==0.3.9
+Flask-Cors==3.0.10
+flask_sqlalchemy==3.0.2
+marshmallow==3.12.1
+marshmallow_sqlalchemy==0.28.1
+psycopg2-binary==2.9.4
+PyJWT==2.1.0
+redis==3.5.3
+requests==2.25.1
+SQLAlchemy==1.4.18
+urllib3==1.26.5
+Werkzeug==2.2.2
+python-dotenv==0.21.0
+gunicorn==20.1.0
+wheel
+google-cloud-storage
+sendgrid
+google-cloud-pubsub
+certifi==2021.5.30
