--- requirementsBK.txt
@@ -1,44 +1,0 @@
-amqp==5.0.6
-aniso8601==9.0.1
-astroid==2.5.6
-billiard==3.6.4.0
-celery==5.2.7
-certifi==2021.5.30
-chardet==4.0.0
-click==8.0.0
-click-didyoumean==0.0.3
-click-plugins==1.1.1
-click-repl==0.2.0
-Flask==2.1.3
-Flask-JWT==0.2.0
-Flask-JWT-Extended==4.2.3
-Flask-RESTful==0.3.9
-Flask-Cors==3.0.10
-Flask-SQLAlchemy==2.5.1
-greenlet==1.1.0
-idna==2.10
-isort==5.8.0
-itsdangerous==2.0.1
-Jinja2==3.0.1
-kombu==5.1.0
-lazy-object-proxy==1.6.0
-MarkupSafe==2.0.1
-marshmallow==3.12.1
-marshmallow-sqlalchemy==0.25.0
-mccabe==0.6.1
-prompt-toolkit==3.0.18
-psycopg2==2.8.6
-PyJWT==2.1.0
-pylint==2.8.3
-pytz==2021.1
-redis==3.5.3
-requests==2.25.1
-six==1.16.0
-SQLAlchemy==1.4.15
-toml==0.10.2
-urllib3==1.26.5
-vine==5.0.0
-wcwidth==0.2.5
-Werkzeug==2.0.1
-wrapt==1.12.1
-cryptography==3.4.7
