--- requirementsNew.txt
@@ -1,7 +1,0 @@
-flask==2.2.2
-flask_cors==3.0.10
-flask_jwt_extended==4.4.4
-flask_restful==0.3.9
-flask_sqlalchemy==3.0.2
-marshmallow_sqlalchemy==0.28.1
-psycopg2==2.9.4
